requirements.txt
torch>=2.0.0
transformers>=4.35.0
diffusers>=0.25.0
accelerate>=0.25.0
safetensors>=0.4.0
Pillow>=10.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0
